feat: change event style
