+++ components/event/card-info.tsx
@@ -1,6 +1,6 @@
 import { EventInfoText } from "./styles/event"
 
-import { EventInfoProps } from "./types/event"
+import { EventInfoProps } from "./types"
 
 export const EventInfo: React.FC<EventInfoProps> = ({ placeName, time }) => {
   return (
