+++ components/event/index.tsx
@@ -1,43 +1,20 @@
-import { useState, useEffect } from "react"
-import axios from "axios"
 import { EventCard } from "./card"
+import { Title } from "./title"
 
 import { Container } from "../../styles/container"
 import { EventContainer } from "./styles/event"
-import { LongerYellowHeadline } from "../../styles/yellow-headline"
 
-import { EventTypeProps } from "./types/event"
+import { useMockEventData } from "./utils/hooks-get-mock-data"
 
-interface EventCardData {
-  img: string
-  name: string
-  place: string
-  time: string
-  price: number
-}
-
-interface ResponseFromServer {
-  data: EventCardData[]
-}
+import { EventTypeProps } from "./types"
 
 export const Events: React.FC<EventTypeProps> = ({ text }) => {
-  const [data, addDataToState] = useState<EventCardData[]>()
-
-  useEffect(() => {
-    axios.get<ResponseFromServer>("/api/event").then(response => {
-      const dataPayload = response.data
-
-      addDataToState(dataPayload.data)
-    })
-  })
+  const data = useMockEventData()
 
   return (
     <Container>
       <EventContainer>
-        <h1 className="recommend">
-          {text}
-          <LongerYellowHeadline />
-        </h1>
+        <Title text={text} />
 
         <div className="cards">
           {data &&
